+++ app/Console/Kernel.php
@@ -21,6 +21,7 @@ class Kernel extends ConsoleKernel
         \Pterodactyl\Console\Commands\ClearTasks::class,
         \Pterodactyl\Console\Commands\ClearServices::class,
         \Pterodactyl\Console\Commands\UpdateEmailSettings::class,
+        \Pterodactyl\Console\Commands\CleanServiceBackup::class,
     ];
 
     /**
@@ -33,5 +34,6 @@ class Kernel extends ConsoleKernel
     {
         $schedule->command('pterodactyl:tasks')->everyMinute()->withoutOverlapping();
         $schedule->command('pterodactyl:tasks:clearlog')->twiceDaily(3, 15);
+        $schedule->command('pterodactyl:cleanservices')->twiceDaily(1, 13);
     }
 }