+++ app/Http/Controllers/Admin/ServiceController.php
@@ -27,6 +27,7 @@ namespace Pterodactyl\Http\Controllers\Admin;
 use DB;
 use Log;
 use Alert;
+use Storage;
 use Pterodactyl\Models;
 use Illuminate\Http\Request;
 use Pterodactyl\Exceptions\DisplayException;
@@ -285,5 +286,40 @@ class ServiceController extends Controller
         }
 
         return redirect()->route('admin.services.option', [$service, $option]);
+    }
+
+    public function getConfiguration(Request $request, $serviceId)
+    {
+        $service = Models\Service::findOrFail($serviceId);
+
+        return view('admin.services.config', [
+            'service' => $service,
+            'contents' => [
+                'json' => Storage::get('services/' . $service->file . '/main.json'),
+                'index' => Storage::get('services/' . $service->file . '/index.js'),
+            ],
+        ]);
+    }
+
+    public function postConfiguration(Request $request, $serviceId)
+    {
+        try {
+            $repo = new ServiceRepository\Service;
+            $repo->updateFile($serviceId, $request->except([
+                '_token',
+            ]));
+
+            return response('', 204);
+        } catch (DisplayException $ex) {
+            return response()->json([
+                'error' => $ex->getMessage(),
+            ], 503);
+        } catch (\Exception $ex) {
+            Log::error($ex);
+
+            return response()->json([
+                'error' => 'An error occured while attempting to save the file.',
+            ], 503);
+        }
     }
 }